Group_06/Source_Code/requirements.txt
Flask
werkzeug
flask-cors
torch
torchvision
transformers
opencv-python
numpy
pandas
scikit-learn
Pillow
grad-cam
